+++ src/types/telegraf-context.ts
@@ -3,6 +3,8 @@ import type { CallbackQuery, Message, Update } from "telegraf/typings/core/types
 
 export type CallbackCtx = NarrowedContext<Context, Update.CallbackQueryUpdate<CallbackQuery.DataQuery>>;
 
+export type EditedMessageCtx = NarrowedContext<Context, Update.EditedMessageUpdate>;
+
 export type MessageCtx = NarrowedContext<Context, Update.MessageUpdate>;
 
 export type MyChatMemberCtx = NarrowedContext<Context, Update.MyChatMemberUpdate>;